--- a/lib/vtshaver.js
+++ b/lib/vtshaver.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var binary = require('node-pre-gyp');
+var binary = require('@mapbox/node-pre-gyp');
 var exists = require('fs').existsSync || require('path').existsSync;
 var path = require('path');
 var binding_path = binary.find(path.resolve(path.join(__dirname,'../package.json')));