--- a/mason-versions.ini
+++ b/mason-versions.ini
@@ -1,6 +1,6 @@
 [headers]
 vtzero=1.0.3
-protozero=1.6.8
+protozero=1.7.0
 gzip-hpp=0.1.0
 [compiled]
 clang++=10.0.0