--- a/mason-versions.ini
+++ b/mason-versions.ini
@@ -1,5 +1,5 @@
 [headers]
-vtzero=1.0.3
+vtzero=1.1.0
 protozero=1.7.0
 gzip-hpp=0.1.0
 [compiled]