--- a/src/filters.cpp
+++ b/src/filters.cpp
@@ -181,8 +181,6 @@ Napi::Value Filters::layers(Napi::CallbackInfo const& info) {
     Napi::HandleScope scope(info.Env());
     auto layers = Napi::Array::New(Env());
     std::uint32_t idx = 0;
-    //auto* filters = info.Holder().Unwrap<Filters>();
-
     for (auto const& lay : filters) {
         layers.Set(idx++, lay.first);
     }