… .travis.yml
… CHANGELOG.md
… docs/benchmarking.md
… docs/extended-tour.md
… docs/howToDebugWithVSCode.md
… docs/image/howtodebug1.png
… docs/image/howtodebug2.png
… docs/image/howtodebug3.png
… docs/npm-and-package-lock.md
… docs/publishing-binaries.md
… package.json
… src/shave.cpp
… src/shave.hpp
… src/vtshaver.cpp
… test/vtshaver.test.js