@axrj 2022-02-17
Local benchmarking results

master
```
ramradi@FVFG62GKQ05P vtshaver % time node bench/bench-batch.js --iterations 50 --concurrency 10
Running benchmark...
Benchmark speed: 14423 runs/s (runs:1500 ms:104 )
Note: pass --mem to track memory usage
Benchmark iterations: 50 concurrency: 10
Success: rate(14423.076923076924 ops/s) > min_rate(1000)
node bench/bench-batch.js --iterations 50 --concurrency 10  1.37s user 0.10s system 180% cpu 0.816 total
```

PR
```
ramradi@FVFG62GKQ05P vtshaver % time node bench/bench-batch.js --iterations 50 --concurrency 10
Running benchmark...
Benchmark speed: 46875 runs/s (runs:1500 ms:32 )
Note: pass --mem to track memory usage
Benchmark iterations: 50 concurrency: 10
Success: rate(46875 ops/s) > min_rate(1000)
node bench/bench-batch.js --iterations 50 --concurrency 10  0.34s user 0.03s system 216% cpu 0.171 total
```
It's much faster probably due to the [async worker](https://github.com/mapbox/vtshaver/pull/55/commits/8f69a70278e8f2de627f981d1d61e17aa20ab75b) implementation